requirements.txt
# ============ WEB FRAMEWORK ============
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# ============ DATABASE ============
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# ============ AUTHENTICATION & SECURITY ============
bcrypt==4.1.1
pyjwt==2.11.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ============ OAUTH ============
google-auth==2.26.2

# ============ DATA VALIDATION ============
pydantic==2.5.3
pydantic-settings==2.1.0

# ============ ENVIRONMENT ============
python-dotenv==1.0.0

# ============ EMAIL ============
secure-smtplib==0.1.1

# ============ UTILITIES ============
aiofiles==23.2.1
python-dateutil==2.8.2
httpx==0.25.2
requests==2.31.0

# ============ LOGGING ============
python-json-logger==2.0.7